Split components into modules
